[#17] Add locking around reads and writes
When reading/writing from/to data objects in a Rados resource, the
librados calls would sometimes hang indefinitely when used with
concurrency. This change adds a lock around the librados read and
write calls to ensure consistency in the Rados IO context being
shared among the threads, which fixes the hang.